refactor: Update FetchIO implementation and remove unused video processing classes
- Replaced the existing FetchIO class with a new implementation in the io directory for better organization.
- Removed the ClipperCore and DecodePipeline classes, which were previously responsible for video frame processing, as they are no longer needed.
- Updated tsconfig to prevent emitting output files during compilation.
- Cleaned up the vite.config by removing commented-out code for clarity.
