+++ src/utils/clipper/io/FetchIO.ts
@@ -1,0 +1,89 @@
+/**
+ * FetchIO 类 - 负责从URL获取指定范围的视频数据
+ * 处理数据块的读取和管理
+ */
+export class FetchIO {
+	/**
+	 * 从URL获取指定范围的ArrayBuffer
+	 * @param url 目标URL
+	 * @param start 起始字节
+	 * @param end 结束字节
+	 * @returns 获取到的ArrayBuffer
+	 */
+	public async fetchBufferRange(
+		url: string,
+		start: number,
+		end: number,
+	): Promise<Response> {
+		const response = fetch(url, {
+			headers: {
+				Range: `bytes=${start}-${end}`,
+			},
+			priority: "low",
+		});
+		return response;
+	}
+
+	/**
+	 * 流式读取视频块
+	 * @param url 视频URL
+	 * @param callback 处理数据的回调函数
+	 * @param options 读取选项
+	 */
+	public async streamChunks(
+		url: string,
+		callback: (buffer: ArrayBuffer, position: number) => Promise<boolean>,
+		options: {
+			initialChunkSize?: number;
+			stepChunkSize?: number;
+			maxSteps?: number;
+		} = {},
+	): Promise<void> {
+		// 初始块大小
+		const initialChunkSize = options.initialChunkSize || 188 * 1536;
+		// 步进块大小
+		const stepChunkSize = options.stepChunkSize || 188 * 1024;
+
+		let currentPosition = 0;
+		let shouldContinue = true;
+
+		// 读取初始块
+		try {
+			const response = await this.fetchBufferRange(
+				url,
+				currentPosition,
+				currentPosition + initialChunkSize - 1,
+			);
+
+			// 调用回调处理数据
+			shouldContinue = await callback(
+				await response.arrayBuffer(),
+				currentPosition,
+			);
+			currentPosition += initialChunkSize;
+
+			// 继续读取后续数据块，直到达到最大步数或回调返回false
+			while (shouldContinue) {
+				const response = await this.fetchBufferRange(
+					url,
+					currentPosition,
+					currentPosition + stepChunkSize - 1,
+				);
+
+				if (response.status === 416) {
+					shouldContinue = false;
+					break;
+				}
+
+				shouldContinue = await callback(
+					await response.arrayBuffer(),
+					currentPosition,
+				);
+				currentPosition += stepChunkSize;
+			}
+		} catch (error) {
+			console.error("流式读取数据出错:", error);
+			throw error;
+		}
+	}
+}
